http://hdl.handle.net/2142/103034
Popularity among peers is important for adolescents. If youth high in social status possess characteristics in line with the values and norms of their culture, they may serve as agents of cultural socialization. Thus, focusing on the United States and China, the goal of this graph is to present the difference in antisocial behavior that contributes to a high level of popularity in the peer system during early adolescence. Our research indicated that popular adolescents in the United States showed higher antisocial behavior than their counterparts in China. Moreover, popular adolescents are more scattered in the American sample than in the Chinese sample, which might imply the more homogeneous norm in terms of attitude to popularity in Chinese Adolescents. This result means that the perception of popularity in peers might link with antisocial behavior in the United States but not in China.


Sources:

The data comes from the Tri-Country Project at the University of Illinois at Urbana-Champaign's Center for Parent-Child Studies.